+++ include/path_tracer/material.h
@@ -1,9 +1,9 @@
 #ifndef path_tracer_material_h
 #define path_tracer_material_h
 
-#include "hitable.h"
-#include "helpers.h"
-#include "texture.h"
+#include "path_tracer/hitable.h"
+#include "path_tracer/helpers.h"
+#include "path_tracer/texture.h"
 
 class Material {
 public:
@@ -13,7 +13,7 @@ public:
         Vec3& attenuation,
         Ray& scattered) const = 0;
     
-    virtual Vec3 emitted(float u, float v, Vec3& p) const {
+    virtual Vec3 emitted(double u, double v, Vec3& p) const {
         return Vec3::zero;
     }
 };
@@ -42,9 +42,9 @@ public:
 class Metal : public Material {
 public:
     Texture* albedo;
-    float fuzz;
+    double fuzz;
     
-    Metal(Texture* tex, float f) : albedo(tex) {
+    Metal(Texture* tex, double f) : albedo(tex) {
         fuzz = f < 1.0f ? f : 1.0f;
     }
     
@@ -64,9 +64,9 @@ public:
 
 class Dielectric : public Material {
 public:
-    float ior;
+    double ior;
 
-    Dielectric(float index_of_refraction) : ior(index_of_refraction) {}
+    Dielectric(double index_of_refraction) : ior(index_of_refraction) {}
 
     virtual bool scatter(
         const Ray& r_in,
@@ -75,13 +75,13 @@ public:
         Ray& scattered) const
     {
         Vec3 outward_normal;
-        float eta; // greek letter η -> η2/η1
+        double eta; // greek letter η -> η2/η1
         attenuation = Vec3(1.0f, 1.0f, 1.0f);
         Vec3 reflected = reflect(r_in.direction, record.normal);
         Vec3 refracted;
-        float reflect_prob;
-        float cosine;
-        float in_dot_n = dot(r_in.direction, record.normal);
+        double reflect_prob;
+        double cosine;
+        double in_dot_n = dot(r_in.direction, record.normal);
     
         if (dot(r_in.direction, record.normal) > 0.0f) {
             outward_normal = -record.normal;
@@ -128,7 +128,7 @@ public:
         return false;
     }
     
-    virtual Vec3 emitted(float u, float v, Vec3& p) const {
+    virtual Vec3 emitted(double u, double v, Vec3& p) const {
         return emit->value(u, v, p);
     }
 };
